pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "clawspec"
version = "0.1.0"
description = "Contract-first testing for OpenClaw skills and agents."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
  { name = "Austin Mao" },
]
dependencies = [
  "PyYAML>=6.0",
  "httpx>=0.27",
  "jsonschema>=4.21",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.11",
]

[project.scripts]
clawspec = "clawspec.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
